+++ activerecord/lib/active_record/integration.rb
@@ -152,6 +152,10 @@ module ActiveRecord
           end
         end
       end
+
+      def collection_cache_key(collection = all, timestamp_column = :updated_at) # :nodoc:
+        collection.compute_cache_key(timestamp_column)
+      end
     end
 
     private
